% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Grouped population data by 5-year age groups (deprecated name)}
\format{
Identical to \link{grouped_pop}: a data frame with 16 rows (age groups)
and 16 variables.
}
\usage{
data
}
\description{
\code{data} is the former name of \link{grouped_pop}, kept so that code written
against earlier versions keeps working. The name shadowed base R's
\code{\link[utils:data]{utils::data()}} function once loaded, which made \code{data(...)} calls in the
same script fail in confusing ways. Use \link{grouped_pop} instead; \code{data} will
be removed in a future release.
}
\seealso{
\link{grouped_pop}, which this duplicates.
}
\keyword{internal}
