% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FertilityModel.R
\name{FertilityModel}
\alias{FertilityModel}
\title{Placeholder Fertility Model for Shiny App}
\description{
R6 Class for a placeholder fertility model. Inherits from DemographicModelBase.
This version has model fitting disabled and uses a heuristic for predictions.
}
\section{Super class}{
\code{\link[DemogAnr:DemographicModelBase]{DemogAnr::DemographicModelBase}} -> \code{FertilityModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{Placeholder for a fitted model object (e.g., from brms). NULL in this version.}

\item{\code{mcmc_settings}}{An instance of McmcSettings.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FertilityModel-new}{\code{FertilityModel$new()}}
\item \href{#method-FertilityModel-fit}{\code{FertilityModel$fit()}}
\item \href{#method-FertilityModel-predict_fertility}{\code{FertilityModel$predict_fertility()}}
\item \href{#method-FertilityModel-clone}{\code{FertilityModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DemogAnr" data-topic="DemographicModelBase" data-id="get_mapping"><a href='../../DemogAnr/html/DemographicModelBase.html#method-DemographicModelBase-get_mapping'><code>DemogAnr::DemographicModelBase$get_mapping()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DemogAnr" data-topic="DemographicModelBase" data-id="prepare_model_data"><a href='../../DemogAnr/html/DemographicModelBase.html#method-DemographicModelBase-prepare_model_data'><code>DemogAnr::DemographicModelBase$prepare_model_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FertilityModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-FertilityModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the FertilityModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FertilityModel$new(data, variable_mapping, mcmc_settings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame or data.table containing demographic data.}

\item{\code{variable_mapping}}{A list mapping standard variable names to actual column names.
Must include a mapping for 'births'.}

\item{\code{mcmc_settings}}{Optional. An existing McmcSettings object. If NULL, a new one
with default settings is created.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FertilityModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-FertilityModel-fit}{}}}
\subsection{Method \code{fit()}}{
Placeholder for model fitting. Fitting is disabled in this version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FertilityModel$fit(formula = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{Optional. A formula for model fitting (ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FertilityModel-predict_fertility"></a>}}
\if{latex}{\out{\hypertarget{method-FertilityModel-predict_fertility}{}}}
\subsection{Method \code{predict_fertility()}}{
Predict fertility (births) using a heuristic approach.
This method does not use a statistical model for prediction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FertilityModel$predict_fertility(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{A data frame or data.table with new data for prediction.
Must contain columns that can be processed by \code{prepare_fertility_data}
to retrieve population information if a population-based heuristic is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of plausible predicted birth numbers, corresponding
to the rows in \code{newdata}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FertilityModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FertilityModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FertilityModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
