% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigrationModel.R
\name{MigrationModel}
\alias{MigrationModel}
\title{Placeholder Migration Model for Shiny App}
\description{
R6 Class for a placeholder migration model. Inherits from DemographicModelBase.
This version has model fitting disabled and uses a heuristic for predictions.
}
\section{Super class}{
\code{\link[DemogAnr:DemographicModelBase]{DemogAnr::DemographicModelBase}} -> \code{MigrationModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{in_model}}{Placeholder for a fitted in-migration model object. NULL in this version.}

\item{\code{out_model}}{Placeholder for a fitted out-migration model object. NULL in this version.}

\item{\code{mcmc_settings}}{An instance of McmcSettings.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MigrationModel-new}{\code{MigrationModel$new()}}
\item \href{#method-MigrationModel-fit}{\code{MigrationModel$fit()}}
\item \href{#method-MigrationModel-predict_migration}{\code{MigrationModel$predict_migration()}}
\item \href{#method-MigrationModel-clone}{\code{MigrationModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DemogAnr" data-topic="DemographicModelBase" data-id="get_mapping"><a href='../../DemogAnr/html/DemographicModelBase.html#method-DemographicModelBase-get_mapping'><code>DemogAnr::DemographicModelBase$get_mapping()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DemogAnr" data-topic="DemographicModelBase" data-id="prepare_model_data"><a href='../../DemogAnr/html/DemographicModelBase.html#method-DemographicModelBase-prepare_model_data'><code>DemogAnr::DemographicModelBase$prepare_model_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the MigrationModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$new(data, variable_mapping, mcmc_settings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame or data.table containing demographic data.}

\item{\code{variable_mapping}}{A list mapping standard variable names to actual column names.
Must include mappings for 'in_migration' and 'out_migration'.}

\item{\code{mcmc_settings}}{Optional. An existing McmcSettings object. If NULL, a new one
with default settings is created.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-fit}{}}}
\subsection{Method \code{fit()}}{
Placeholder for model fitting. Fitting is disabled in this version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$fit(in_formula = NULL, out_formula = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_formula}}{Optional. Formula for in-migration model (ignored).}

\item{\code{out_formula}}{Optional. Formula for out-migration model (ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-predict_migration"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-predict_migration}{}}}
\subsection{Method \code{predict_migration()}}{
Predict in-migration and out-migration counts using a heuristic approach.
This method does not use a statistical model for prediction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$predict_migration(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{A data frame or data.table with new data for prediction.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing two named numeric vectors:
\code{in_migration}: plausible predicted in-migration counts.
\code{out_migration}: plausible predicted out-migration counts.
Each vector corresponds to the rows in \code{newdata}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
