% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityModel.R
\name{MortalityModel}
\alias{MortalityModel}
\title{Placeholder Mortality Model for Shiny App}
\description{
R6 Class for a placeholder mortality model. Inherits from DemographicModelBase.
This version has model fitting disabled and uses a heuristic for predictions.
}
\section{Super class}{
\code{\link[DemogAnr:DemographicModelBase]{DemogAnr::DemographicModelBase}} -> \code{MortalityModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{Placeholder for a fitted model object (e.g., from brms). NULL in this version.}

\item{\code{mcmc_settings}}{An instance of McmcSettings.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MortalityModel-new}{\code{MortalityModel$new()}}
\item \href{#method-MortalityModel-fit}{\code{MortalityModel$fit()}}
\item \href{#method-MortalityModel-predict_mortality}{\code{MortalityModel$predict_mortality()}}
\item \href{#method-MortalityModel-clone}{\code{MortalityModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DemogAnr" data-topic="DemographicModelBase" data-id="get_mapping"><a href='../../DemogAnr/html/DemographicModelBase.html#method-DemographicModelBase-get_mapping'><code>DemogAnr::DemographicModelBase$get_mapping()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DemogAnr" data-topic="DemographicModelBase" data-id="prepare_model_data"><a href='../../DemogAnr/html/DemographicModelBase.html#method-DemographicModelBase-prepare_model_data'><code>DemogAnr::DemographicModelBase$prepare_model_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the MortalityModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityModel$new(data, variable_mapping, mcmc_settings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame or data.table containing demographic data.}

\item{\code{variable_mapping}}{A list mapping standard variable names to actual column names.
Must include a mapping for 'deaths'.}

\item{\code{mcmc_settings}}{Optional. An existing McmcSettings object. If NULL, a new one
with default settings is created.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityModel-fit}{}}}
\subsection{Method \code{fit()}}{
Placeholder for model fitting. Fitting is disabled in this version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityModel$fit(formula = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{Optional. A formula for model fitting (ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityModel-predict_mortality"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityModel-predict_mortality}{}}}
\subsection{Method \code{predict_mortality()}}{
Predict mortality rates using a heuristic approach.
This method does not use a statistical model for prediction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityModel$predict_mortality(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{A data frame or data.table with new data for prediction.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of plausible predicted mortality \emph{rates} (deaths per person),
corresponding to the rows in \code{newdata}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
