% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brass_logits.R
\name{brass_logit}
\alias{brass_logit}
\title{Brass Logit Model}
\usage{
brass_logit(
  data,
  qx_col,
  age_col,
  standard,
  standards_data = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame containing the observed probabilities (nqx) and ages.}

\item{qx_col}{Column name for nqx in the data frame.}

\item{age_col}{Column name for age in the data frame.}

\item{standard}{Standard to use ("African" or "GeneralUN").}

\item{standards_data}{Standards dataset (default: included "standards").}

\item{verbose}{Logical. If \code{TRUE}, prints detailed status messages to the console.}
}
\value{
A list containing the model, coefficients, and modified data with predicted nqx.
}
\description{
Brass Logit Model
}
\examples{
observed_data <- data.frame(
  age = c(0, 1, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50),
  qx = c(0.1, 0.05, 0.02, NA, 0.005, NA, 0.002, 0.0015, NA, 0.0008, 0.0006, 0.0005)
)

model<-brass_logit(
  data = observed_data,
  qx_col = "qx",
  age_col = "age",
  standard = "African"
)

lifetable <- model$data

lifetable

lifetable_nqx(lifetable, age="age", nqx = "predicted_qx")
}
\references{
Brass, W. (1971). On the scale of mortality. \emph{Biological Aspects of Demography}, 69-110. London: Taylor & Francis.

Moultrie, T., Dorrington, R., Hill, A., Hill, K., Timæus, I., & Zaba, B. (2013). \emph{Tools for Demographic Estimation}. Paris: International Union for the Scientific Study of Population (IUSSP). (Section: Relational Logit Models).

Newell, C. (1988). \emph{Methods and Models in Demography}. (Chapter 10: Model Life Tables).
}
