% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDR_ASDR.R
\name{dem.cdr}
\alias{dem.cdr}
\title{Calculate CDR and ASDR}
\usage{
dem.cdr(data, type, age_col = NULL, population_col, deaths_col)
}
\arguments{
\item{data}{A data frame containing demographic data.}

\item{type}{A string or vector specifying the type(s) of death rate calculation.
Options: "CDR", "ASDR". Use "all" to calculate both.}

\item{age_col}{The column name for age groups (used for "ASDR").}

\item{population_col}{The column name for total population.}

\item{deaths_col}{The column name for total deaths.}
}
\value{
A list of named outputs for the requested death rate calculations.
}
\description{
This function calculates the Crude Death Rate (CDR) and Age-Specific Death Rates (ASDR).
}
\examples{
data <- data.frame(
  age = c("0-4", "5-9", "10-14"),
  population = c(10000, 12000, 11000),
  deaths = c(100, 50, 30)
)
dem.cdr(data, type = "all", age_col = "age",
                    population_col = "population", deaths_col = "deaths")
}
\references{
Preston, S. H., Heuveline, P., & Guillot, M. (2001). \emph{Demography: Measuring and Modeling Population Processes}. Oxford: Blackwell Publishers. (Chapters 2 and 5)

Newell, C. (1988). \emph{Methods and Models in Demography}. New York: Guilford Press. (Chapters 4 and 6)

Siegel, J. S., & Swanson, D. A. (Eds.). (2004). \emph{The Methods and Materials of Demography} (2nd ed.). Emerald Group Publishing. (Chapters 14 for Maternal Mortality)
}
