% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dem-package.R
\docType{package}
\name{DemogAnr-package}
\alias{DemogAnr}
\alias{DemogAnr-package}
\title{DemogAnr: Demographic Analysis}
\description{
Tool for demographic modeling and analysis, combining stochastic simulation-based projections with classic demographic methods. Provides utilities for disaggregating population data using Karup-King interpolation, fitting Brass relational logit models, calculating fertility and mortality metrics, building life tables, and performing hierarchical subnational population projections. Relational life table estimation and interpolation methods are described in Brass (1975) "Methods for Estimating Fertility and Mortality from Limited and Defective Data", Preston et al. (2001, ISBN:978-0631226161) "Demography: Measuring and Modeling Population Processes", Siegel and Swanson (2004, ISBN:978-0126419559) "The Methods and Materials of Demography", and Raftery et al. (2012) "Bayesian probabilistic population projections for all countries" \doi{10.1073/pnas.1211452109}.
}
\author{
\strong{Maintainer}: Edward Owusu Manu \email{edward.manu@statsghana.gov.gh}

}
\keyword{internal}
