% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetables.R
\name{Life Table Computations}
\alias{Life Table Computations}
\alias{lifetable}
\title{lifetable}
\usage{
lifetable(data, age = "Age", nMx = NULL, pop = NULL, Dx = NULL)
}
\arguments{
\item{data}{A data frame containing age-specific mortality probabilities.}

\item{age}{Numeric vector: age groups to be presented in point format, e.g., 0, 1, 5, 10, ...}

\item{nMx}{Numeric vector (optional): the observed mortality rates in the population. If not provided, \code{pop} and \code{Dx} must be given to compute it.}

\item{pop}{Numeric vector (optional): total population in each age group. Used to compute \code{nMx} if not given.}

\item{Dx}{Numeric vector (optional): deaths occurred in each age group. Used to compute \code{nMx} if not given.}
}
\value{
A list containing:
\describe{
\item{\code{metrics}}{A list of the life expectncy at birth and the total deaths recorded}
\item{\code{lifetable}}{A dataframe of the complete life table.}
}
}
\description{
This function computes a complete life table based on given mortality rates or raw mortality data.
It takes a dataframe with the necessary inputs and computes standard life table metrics.
}
\examples{
#Example without moratlity rate
data(gphc2010)
lifetable(data=gphc2010, age = "Age", pop = "Pop", Dx = "Deaths")


}
